Feature/attachment manage

+++ EShopSolution.Data/Entities/ProductImage.cs
@@ -1,7 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace EShopSolution.Data.Entities
 {
@@ -14,7 +11,7 @@ namespace EShopSolution.Data.Entities
         public bool IsDefault { get; set; }
         public DateTime? DateCreated { get; set; }
         public int SortOrder { get; set; }
-        public int FileSize { get; set; }
+        public long FileSize { get; set; }
         public Product Product { get; set; }
     }
 }
